+++ pkg/remotewriteutils/error.go
@@ -1,0 +1,11 @@
+package remotewriteutils
+
+import "github.com/giantswarm/microerror"
+
+var wrongTypeError = &microerror.Error{
+	Kind: "wrongTypeError",
+}
+
+var errorFetchingPrometheus = &microerror.Error{
+	Kind: "errorFetchingPrometheus",
+}
