+++ service/controller/resource/remotewritesecret/create.go
@@ -6,20 +6,24 @@ import (
 
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/operatorkit/v7/pkg/controller/context/resourcecanceledcontext"
+	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
+
+	"github.com/giantswarm/prometheus-meta-operator/pkg/remotewriteutils"
 )
 
 func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 	r.logger.Debugf(ctx, "ensuring prometheus remotewrite secret")
 	{
-		remoteWrite, err := ToRemoteWrite(obj)
+		remoteWrite, err := remotewriteutils.ToRemoteWrite(obj)
 		if err != nil {
 			return microerror.Mask(err)
 		}
 
 		// fetch current prometheus using the selector provided in remoteWrite resource.
-		prometheusList, err := fetchPrometheusList(ctx, r, remoteWrite)
+		prometheusList, err := remotewriteutils.FetchPrometheusList(ctx, toResourceWrapper(r), remoteWrite)
 		if err != nil {
 			return microerror.Mask(err)
 		}
@@ -31,24 +35,46 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 
 		for _, current := range prometheusList.Items {
 
+			installedSecrets := make([]corev1.Secret, 0)
 			// Loop over remote write secrets
 			for _, item := range remoteWrite.Spec.Secrets {
 				desired := r.ensureRemoteWriteSecret(item, remoteWrite.ObjectMeta, current.GetNamespace())
 				secret, err := r.k8sClient.K8sClient().CoreV1().Secrets(current.GetNamespace()).Get(ctx, item.Name, metav1.GetOptions{})
 				if apierrors.IsNotFound(err) {
 					r.logger.Debugf(ctx, fmt.Sprintf("creating Secret %#q in namespace %#q", desired.Name, desired.Namespace))
-
-					//if err := r.k8sClient.K8sClient().SetControllerReference(p, &desiredSa, r.Scheme); err != nil {
-					//	return err
-					//}
 					secret, err = r.k8sClient.K8sClient().CoreV1().Secrets(current.GetNamespace()).Create(ctx, &desired, metav1.CreateOptions{})
 				}
 				if err != nil {
 					return microerror.Mask(err)
 				}
+				installedSecrets = append(installedSecrets, *secret)
 				r.logger.Debugf(ctx, fmt.Sprintf("Secret %#q in namespace %#q created", secret.Name, secret.Namespace))
 
 			}
+
+			/*
+			  Cleanup deleted secrets from RemoteWrite CR
+			*/
+			l := labels.SelectorFromSet(labels.Set(map[string]string{label: Name}))
+			options := metav1.ListOptions{
+				LabelSelector: l.String(),
+			}
+			secrets, err := r.k8sClient.K8sClient().CoreV1().Secrets(current.GetNamespace()).List(ctx, options)
+			if err != nil {
+				return microerror.Mask(err)
+			}
+			if secrets != nil && len(secrets.Items) > 0 {
+				for _, secret := range secrets.Items {
+					// delete secret if it doesn't exist in the remotewrite secrets field
+					if !secretInstalled(secret, installedSecrets) {
+						err := r.performDelete(ctx, secret.GetName(), secret.GetNamespace())
+						if err != nil {
+							return microerror.Mask(err)
+						}
+					}
+				}
+			}
+
 		}
 
 	}
