+++ service/controller/resource/remotewritesecret/delete.go
@@ -5,19 +5,20 @@ import (
 
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/operatorkit/v7/pkg/controller/context/resourcecanceledcontext"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	"github.com/giantswarm/prometheus-meta-operator/pkg/remotewriteutils"
 )
 
 func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
 	r.logger.Debugf(ctx, "deleting prometheus remoteWrite secret")
 	{
-		remoteWrite, err := ToRemoteWrite(obj)
+		remoteWrite, err := remotewriteutils.ToRemoteWrite(obj)
 		if err != nil {
 			return microerror.Mask(err)
 		}
 
 		// fetch current prometheus using the selector provided in remoteWrite resource.
-		prometheusList, err := fetchPrometheusList(ctx, r, remoteWrite)
+		prometheusList, err := remotewriteutils.FetchPrometheusList(ctx, toResourceWrapper(r), remoteWrite)
 		if err != nil {
 			return microerror.Mask(err)
 		}
@@ -31,7 +32,7 @@ func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
 
 			// Loop over remote write secrets
 			for _, item := range remoteWrite.Spec.Secrets {
-				err := r.k8sClient.K8sClient().CoreV1().Secrets(current.GetNamespace()).Delete(ctx, item.Name, metav1.DeleteOptions{})
+				err := r.performDelete(ctx, item.Name, current.GetNamespace())
 				if err != nil {
 					return microerror.Mask(err)
 				}
