+++ service/controller/resource/remotewritesecret/error.go
@@ -1,11 +1,1 @@
 package remotewritesecret
-
-import "github.com/giantswarm/microerror"
-
-var wrongTypeError = &microerror.Error{
-	Kind: "wrongTypeError",
-}
-
-var errorFetchingPrometheus = &microerror.Error{
-	Kind: "errorFetchingPrometheus",
-}
