+++ service/controller/resource/remotewritesecret/resource.go
@@ -4,18 +4,18 @@ import (
 	"context"
 
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
-	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
-	promv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	promclient "github.com/prometheus-operator/prometheus-operator/pkg/client/versioned"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	pmov1alpha1 "github.com/giantswarm/prometheus-meta-operator/api/v1alpha1"
+	"github.com/giantswarm/prometheus-meta-operator/pkg/remotewriteutils"
 )
 
 const (
-	Name = "remotewrite"
+	Name  = "remotewrite"
+	label = "monitoring.giantswarm.io/remotewrite"
 )
 
 type Config struct {
@@ -44,40 +44,44 @@ func (r *Resource) Name() string {
 	return Name
 }
 
-func ToRemoteWrite(obj interface{}) (*pmov1alpha1.RemoteWrite, error) {
-	remotewrite, ok := obj.(*pmov1alpha1.RemoteWrite)
-	if !ok {
-		return nil, microerror.Maskf(wrongTypeError, "'%T' is not a 'pmov1alpha1.RemoteWrite'", obj)
-	}
-
-	return remotewrite, nil
-}
-
-func fetchPrometheusList(ctx context.Context, r *Resource, rw *pmov1alpha1.RemoteWrite) (*promv1.PrometheusList, error) {
-	selector, err := metav1.LabelSelectorAsSelector(&rw.Spec.ClusterSelector)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-	prometheusList, err := r.prometheusClient.
-		MonitoringV1().
-		Prometheuses(metav1.NamespaceAll).
-		List(ctx, metav1.ListOptions{LabelSelector: selector.String()})
-	if err != nil {
-		return nil, microerror.Maskf(errorFetchingPrometheus, "Could not fetch Prometheus with label selector %#q", rw.Spec.ClusterSelector.String())
-	}
-
-	return prometheusList, nil
-}
-
 func (r *Resource) ensureRemoteWriteSecret(scSpec pmov1alpha1.RemoteWriteSecretSpec, meta metav1.ObjectMeta, ns string) corev1.Secret {
+
+	labels := meta.GetLabels()
+	if labels != nil {
+		labels[label] = Name
+	} else {
+		labels = map[string]string{label: Name}
+	}
 
 	return corev1.Secret{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:        scSpec.Name,
 			Namespace:   ns,
-			Labels:      meta.GetLabels(),
+			Labels:      labels,
 			Annotations: meta.GetAnnotations(),
 		},
 		Data: scSpec.Data,
 	}
+}
+
+func (r *Resource) performDelete(ctx context.Context, name, ns string) error {
+	return r.k8sClient.K8sClient().CoreV1().Secrets(ns).Delete(ctx, name, metav1.DeleteOptions{})
+}
+
+func toResourceWrapper(r *Resource) *remotewriteutils.ResourceWrapper {
+	return &remotewriteutils.ResourceWrapper{
+		K8sClient:        r.k8sClient,
+		Logger:           r.logger,
+		PrometheusClient: r.prometheusClient,
+	}
+}
+
+func secretInstalled(sc corev1.Secret, list []corev1.Secret) bool {
+	for _, item := range list {
+		if item.GetNamespace() == sc.GetNamespace() &&
+			item.GetName() == sc.GetName() {
+			return true
+		}
+	}
+	return false
 }
